Worked throuhg refactor doc and implemented suggestions
