+++ web/src/Routes.tsx
@@ -2,12 +2,12 @@ import * as React from 'react'
 
 import SwaggerUI from 'swagger-ui-react'
 import { Routes as Switch, Route } from 'react-router-dom'
-import DocumentationArticle from './docs/Documentation'
-import SampleView from './infoViews/SampleView'
-import FamilyView from './infoViews/FamilyView'
-import ProjectSummary from './project/ProjectSummary'
-import ProjectsAdmin from './admin/ProjectsAdmin'
-import GraphQL from './graphQL'
+import DocumentationArticle from './pages/docs/Documentation'
+import SampleView from './pages/sample/SampleView'
+import FamilyView from './pages/family/FamilyView'
+import ProjectSummary from './pages/project/ProjectSummary'
+import ProjectsAdmin from './pages/admin/ProjectsAdmin'
+import ErrorBoundary from './shared/utilities/errorBoundary'
 
 const Routes: React.FunctionComponent = () => (
     <Switch>
@@ -19,20 +19,62 @@ const Routes: React.FunctionComponent = () => (
         <Route path="/swagger" element={<SwaggerUI url="/openapi.json" />} />
 
         <Route path="/project/">
-            <Route path="" element={<ProjectSummary />} />
-            <Route path=":projectName" element={<ProjectSummary />} />
-            <Route path=":projectName/:page" element={<ProjectSummary />} />
+            <Route
+                path=""
+                element={
+                    <ErrorBoundary>
+                        <ProjectSummary />
+                    </ErrorBoundary>
+                }
+            />
+            <Route
+                path=":projectName"
+                element={
+                    <ErrorBoundary>
+                        <ProjectSummary />
+                    </ErrorBoundary>
+                }
+            />
+            <Route
+                path=":projectName/:page"
+                element={
+                    <ErrorBoundary>
+                        <ProjectSummary />
+                    </ErrorBoundary>
+                }
+            />
         </Route>
 
         <Route path="admin" element={<ProjectsAdmin />} />
 
         <Route path="/" element={<DocumentationArticle articleid="index" />} />
 
-        <Route path="sample/:sampleName" element={<SampleView />} />
-        <Route path="project/:projectName/family/:familyID" element={<FamilyView />} />
-        <Route path="project/:projectName/participant/:participantName" element={<SampleView />} />
+        <Route
+            path="sample/:sampleName"
+            element={
+                <ErrorBoundary>
+                    <SampleView />
+                </ErrorBoundary>
+            }
+        />
 
-        <Route path="graphqltest" element={<GraphQL />} />
+        <Route
+            path="/family/:familyID"
+            element={
+                <ErrorBoundary>
+                    <FamilyView />
+                </ErrorBoundary>
+            }
+        />
+
+        <Route
+            path="project/:projectName/participant/:participantName"
+            element={
+                <ErrorBoundary>
+                    <SampleView />
+                </ErrorBoundary>
+            }
+        />
     </Switch>
 )
