+++ web/src/index.css
@@ -89,15 +89,7 @@ blockquote {
 }
 
 @media screen and (min-width: 1200px) {
-    .detailedInfo {
-        max-width: 1200px;
-        margin-left: auto;
-        margin-right: auto;
-    }
-}
-
-@media screen and (min-width: 1200px) {
-    .familyView {
+    .dataStyle {
         max-width: 1200px;
         margin-left: auto;
         margin-right: auto;
