+++ web/src/pages/docs/Documentation.tsx
@@ -44,7 +44,7 @@ const DocumentationArticle: React.FunctionComponent<IDocumentationArticleProps> 
                 // eg: /static/media/README.e3454f58.md
 
                 // eslint-disable-next-line
-                const file = require(`../static/sm_docs/${filename}`)
+                const file = require(`../../static/sm_docs/${filename}`)
                 const response = await fetch(file.default)
                 let text_ = await response.text()
                 // Do some URL rewriting, because the openapi generator
