+++ web/src/pages/project/ProjectGrid.tsx
@@ -1,0 +1,138 @@
+import * as React from 'react'
+import _ from 'lodash'
+import { Table } from 'semantic-ui-react'
+
+import SampleLink from '../../shared/components/links/SampleLink'
+import FamilyLink from '../../shared/components/links/FamilyLink'
+import sanitiseValue from '../../shared/utilities/sanitiseValue'
+import { ProjectSummaryResponse } from '../../sm-api/api'
+import MuckError from '../../shared/components/MuckError'
+
+interface ProjectGridProps {
+    summary?: ProjectSummaryResponse
+    projectName?: string
+    error?: string
+}
+
+const ProjectGrid: React.FunctionComponent<ProjectGridProps> = ({
+    summary,
+    projectName,
+    error,
+}) => {
+    if (!projectName) {
+        return (
+            <p>
+                <em>Please select a project</em>
+            </p>
+        )
+    }
+    if (error) {
+        return (
+            <p>
+                <em>An error occurred when fetching samples: {error}</em>
+            </p>
+        )
+    }
+    if (!summary) {
+        return <p>Loading...</p>
+    }
+    if (summary.participants.length === 0) {
+        return <MuckError message={`Ah Muck, there aren't any samples in this project`} />
+    }
+    const headers = [
+        'Family ID',
+        ...summary.participant_keys.map((field) => field[1]),
+        ...summary.sample_keys.map((field) => field[1]),
+        ...summary.sequence_keys.map((field) => `sequence.${field[1]}`),
+    ]
+
+    return (
+        <Table celled>
+            <Table.Header>
+                <Table.Row>
+                    {headers.map((k, i) => (
+                        <Table.HeaderCell key={`${k}-${i}`}>{k}</Table.HeaderCell>
+                    ))}
+                </Table.Row>
+            </Table.Header>
+            <Table.Body>
+                {summary.participants.map((p, pidx) =>
+                    p.samples.map((s, sidx) => {
+                        const backgroundColor =
+                            pidx % 2 === 0 ? 'white' : 'var(--bs-table-striped-bg)'
+                        const lengthOfParticipant = p.samples
+                            .map((s_) => s_.sequences.length)
+                            .reduce((a, b) => a + b, 0)
+                        return s.sequences.map((seq, seqidx) => {
+                            const isFirstOfGroup = sidx === 0 && seqidx === 0
+                            return (
+                                <Table.Row key={`${p.external_id}-${s.id}-${seq.id}`}>
+                                    {isFirstOfGroup && (
+                                        <Table.Cell
+                                            style={{ backgroundColor }}
+                                            rowSpan={lengthOfParticipant}
+                                        >
+                                            {
+                                                <FamilyLink
+                                                    id={p.families.map((f) => f.id).join(', ')}
+                                                    projectName={projectName}
+                                                >
+                                                    {p.families
+                                                        .map((f) => f.external_id)
+                                                        .join(', ')}
+                                                </FamilyLink>
+                                            }
+                                        </Table.Cell>
+                                    )}
+                                    {isFirstOfGroup &&
+                                        summary.participant_keys.map(([k]) => (
+                                            <Table.Cell
+                                                style={{
+                                                    backgroundColor,
+                                                }}
+                                                key={`${p.id}participant.${k}`}
+                                                rowSpan={lengthOfParticipant}
+                                            >
+                                                {sanitiseValue(_.get(p, k))}
+                                            </Table.Cell>
+                                        ))}
+                                    {seqidx === 0 &&
+                                        summary.sample_keys.map(([k]) => (
+                                            <Table.Cell
+                                                style={{
+                                                    backgroundColor,
+                                                }}
+                                                key={`${s.id}sample.${k}`}
+                                                rowSpan={s.sequences.length}
+                                            >
+                                                {k === 'external_id' || k === 'id' ? (
+                                                    <SampleLink id={s.id} projectName={projectName}>
+                                                        {sanitiseValue(_.get(s, k))}
+                                                    </SampleLink>
+                                                ) : (
+                                                    sanitiseValue(_.get(s, k))
+                                                )}
+                                            </Table.Cell>
+                                        ))}
+                                    {seq &&
+                                        summary.sequence_keys.map(([k]) => (
+                                            <Table.Cell
+                                                style={{
+                                                    backgroundColor,
+                                                }}
+                                                key={`${s.id}sequence.${k}`}
+                                            >
+                                                {sanitiseValue(_.get(seq, k))}
+                                            </Table.Cell>
+                                        ))}
+                                </Table.Row>
+                            )
+                        })
+                    })
+                )}
+            </Table.Body>
+        </Table>
+    )
+}
+
+export default ProjectGrid
