--- web/src/project/ProjectSummary.tsx
@@ -1,457 +1,0 @@
-import * as React from 'react'
-import { useParams, useNavigate, useSearchParams } from 'react-router-dom'
-import _ from 'lodash'
-
-import { Table, Button, Dropdown } from 'semantic-ui-react'
-import ProjectSelector from './ProjectSelector'
-import { WebApi, ProjectSummaryResponse } from '../sm-api/api'
-
-import { SampleLink, FamilyLink } from '../renders/Links'
-import MuckTheDuck from '../renders/MuckTheDuck'
-
-const PAGE_SIZES = [20, 40, 100, 1000]
-
-const REPORT_PREFIX = 'https://main-web.populationgenomics.org.au/'
-
-const REPORT_TYPES = {
-    'WGS Cram': '/qc/cram/multiqc.html',
-    'WGS GVCF': '/qc/gvcf/multiqc.html',
-    'WGS FASTQC': 'qc/fastqc/multiqc.html',
-    'Exome Cram': 'exome/qc/cram/multiqc.html',
-    'Exome GVCF': 'exome/qc/gvcf/multiqc.html',
-    'Exome FASTQC': 'exome/qc/fastqc/multiqc.html',
-}
-
-const sanitiseValue = (value: any) => {
-    const tvalue = typeof value
-    if (tvalue === 'string' || tvalue === 'number') return value
-    if (tvalue === 'boolean') return value ? 'true' : 'false'
-    if (value === undefined || value === null) return ''
-    return JSON.stringify(value)
-}
-
-const ProjectSummary: React.FunctionComponent = () => {
-    const navigate = useNavigate()
-
-    const { projectName, page } = useParams()
-
-    const [searchParams] = useSearchParams()
-    const pageSize = searchParams.get('size') || 20
-    let directToken = 0
-    if (page && pageSize) {
-        directToken = +page * +pageSize - +pageSize
-    }
-
-    const validPages = !!(page && +page && pageSize && +pageSize && PAGE_SIZES.includes(+pageSize))
-
-    const [summary, setSummary] = React.useState<ProjectSummaryResponse | undefined>()
-    const [pageNumber, setPageNumber] = React.useState<number>(validPages ? +page : 1)
-    const [isLoading, setIsLoading] = React.useState<boolean>(false)
-    const [error, setError] = React.useState<string | undefined>()
-    const [pageLimit, _setPageLimit] = React.useState<number>(
-        validPages ? +pageSize : PAGE_SIZES[0]
-    )
-
-    const handleOnClick = React.useCallback(
-        (p) => {
-            navigate(`/project/${projectName}/${p}?size=${pageLimit}`)
-            setPageNumber(p)
-        },
-        [navigate, pageLimit, projectName]
-    )
-
-    const getProjectSummary = React.useCallback(
-        async (token: any) => {
-            if (!projectName) {
-                setSummary(undefined)
-                return
-            }
-            const sanitisedToken = token || undefined
-            setError(undefined)
-            setIsLoading(true)
-            new WebApi()
-                .getProjectSummary(projectName, pageLimit, sanitisedToken)
-                .then((resp) => {
-                    setIsLoading(false)
-                    setSummary(resp.data)
-                })
-                .catch((er) => {
-                    setIsLoading(false)
-                    setError(er.message)
-                })
-        },
-        [projectName, pageLimit]
-    )
-
-    const setPageLimit = React.useCallback(
-        (e: React.SyntheticEvent<HTMLElement>, { value }) => {
-            navigate(`/project/${projectName}/1?size=${parseInt(value, 10)}`)
-            _setPageLimit(parseInt(value, 10))
-            setPageNumber(1)
-        },
-        [projectName, navigate]
-    )
-
-    const _updateProjectSummary = React.useCallback(() => {
-        getProjectSummary(directToken)
-    }, [getProjectSummary, directToken])
-
-    // retrigger if project changes, or pageLimit changes
-    React.useEffect(_updateProjectSummary, [
-        projectName,
-        pageLimit,
-        pageNumber,
-        _updateProjectSummary,
-    ])
-
-    const totalPageNumbers = Math.ceil((summary?.total_samples || 0) / pageLimit)
-
-    const pageOptions = (
-        <>
-            {pageNumber > 1 && (
-                <Button
-                    style={{ marginLeft: '10px' }}
-                    disabled={isLoading}
-                    onClick={() => {
-                        handleOnClick(pageNumber - 1)
-                    }}
-                >
-                    Previous
-                </Button>
-            )}
-            {!!summary?.total_samples && (
-                <span style={{ padding: '8px 10px 0 10px' }}>
-                    Page {pageNumber} / {totalPageNumbers} ({summary?.total_samples} samples)
-                </span>
-            )}
-            {pageNumber < totalPageNumbers && (
-                <Button
-                    disabled={isLoading}
-                    onClick={() => {
-                        handleOnClick(pageNumber + 1)
-                    }}
-                >
-                    Next
-                </Button>
-            )}
-        </>
-    )
-
-    const renderGrid = () => {
-        if (!projectName) {
-            return (
-                <p>
-                    <em>Please select a project</em>
-                </p>
-            )
-        }
-        if (error) {
-            return (
-                <p>
-                    <em>An error occurred when fetching samples: {error}</em>
-                </p>
-            )
-        }
-        if (!summary) {
-            return <p>Loading...</p>
-        }
-        if (summary.participants.length === 0) {
-            return (
-                <p>
-                    <em>Ah Muck, there aren&apos;t any samples in this project</em>
-                    <MuckTheDuck height={28} style={{ transform: 'scaleY(-1)' }} />
-                </p>
-            )
-        }
-        const headers = [
-            'Family ID',
-            ...summary.participant_keys.map((field) => field[1]),
-            ...summary.sample_keys.map((field) => field[1]),
-            ...summary.sequence_keys.map((field) => `sequence.${field[1]}`),
-        ]
-
-        return (
-            <Table celled>
-                <Table.Header>
-                    <Table.Row>
-                        {headers.map((k, i) => (
-                            <Table.HeaderCell key={`${k}-${i}`}>{k}</Table.HeaderCell>
-                        ))}
-                    </Table.Row>
-                </Table.Header>
-                <Table.Body>
-                    {summary.participants.map((p, pidx) =>
-                        p.samples.map((s, sidx) => {
-                            const backgroundColor =
-                                pidx % 2 === 0 ? 'white' : 'var(--bs-table-striped-bg)'
-                            const lengthOfParticipant = p.samples
-                                .map((s_) => s_.sequences.length)
-                                .reduce((a, b) => a + b, 0)
-                            return s.sequences.map((seq, seqidx) => {
-                                const isFirstOfGroup = sidx === 0 && seqidx === 0
-                                return (
-                                    <Table.Row key={`${p.external_id}-${s.id}-${seq.id}`}>
-                                        {isFirstOfGroup && (
-                                            <Table.Cell
-                                                style={{ backgroundColor }}
-                                                rowSpan={lengthOfParticipant}
-                                            >
-                                                {
-                                                    <FamilyLink
-                                                        id={p.families.map((f) => f.id).join(', ')}
-                                                        projectName={projectName}
-                                                    >
-                                                        {p.families
-                                                            .map((f) => f.external_id)
-                                                            .join(', ')}
-                                                    </FamilyLink>
-                                                }
-                                            </Table.Cell>
-                                        )}
-                                        {isFirstOfGroup &&
-                                            summary.participant_keys.map(([k]) => (
-                                                <Table.Cell
-                                                    style={{
-                                                        backgroundColor,
-                                                    }}
-                                                    key={`${p.id}participant.${k}`}
-                                                    rowSpan={lengthOfParticipant}
-                                                >
-                                                    {sanitiseValue(_.get(p, k))}
-                                                </Table.Cell>
-                                            ))}
-                                        {seqidx === 0 &&
-                                            summary.sample_keys.map(([k]) => (
-                                                <Table.Cell
-                                                    style={{
-                                                        backgroundColor,
-                                                    }}
-                                                    key={`${s.id}sample.${k}`}
-                                                    rowSpan={s.sequences.length}
-                                                >
-                                                    {k === 'external_id' || k === 'id' ? (
-                                                        <SampleLink
-                                                            id={s.id}
-                                                            projectName={projectName}
-                                                        >
-                                                            {sanitiseValue(_.get(s, k))}
-                                                        </SampleLink>
-                                                    ) : (
-                                                        sanitiseValue(_.get(s, k))
-                                                    )}
-                                                </Table.Cell>
-                                            ))}
-                                        {seq &&
-                                            summary.sequence_keys.map(([k]) => (
-                                                <Table.Cell
-                                                    style={{
-                                                        backgroundColor,
-                                                    }}
-                                                    key={`${s.id}sequence.${k}`}
-                                                >
-                                                    {sanitiseValue(_.get(seq, k))}
-                                                </Table.Cell>
-                                            ))}
-                                    </Table.Row>
-                                )
-                            })
-                        })
-                    )}
-                </Table.Body>
-            </Table>
-        )
-    }
-
-    const titleCase = (s: string) => s[0].toUpperCase() + s.slice(1).toLowerCase()
-
-    const batchTable = () => {
-        const seqTypes = Object.keys(summary?.cram_seqr_stats ?? {})
-        const batchEntries = Object.entries(summary?.batch_sequence_stats ?? {})
-
-        if (!seqTypes.length || !batchEntries.length) {
-            return <></>
-        }
-
-        return (
-            <Table celled compact>
-                <Table.Header>
-                    <Table.Row>
-                        <Table.HeaderCell>Batch</Table.HeaderCell>
-                        {seqTypes.map((item) => (
-                            <Table.HeaderCell key={`header-${item}-${projectName}`}>
-                                {titleCase(item)}
-                            </Table.HeaderCell>
-                        ))}
-                        <Table.HeaderCell>Total</Table.HeaderCell>
-                    </Table.Row>
-                </Table.Header>
-
-                <Table.Body>
-                    {batchEntries
-                        .sort((a, b) => {
-                            if (a[0] === b[0]) {
-                                return 0
-                            }
-                            if (a[0] === 'no-batch') {
-                                return 1
-                            }
-                            if (b[0] === 'no-batch') {
-                                return -1
-                            }
-                            return a[0].localeCompare(b[0], undefined, {
-                                numeric: true,
-                            })
-                        })
-                        .map(([key, value]) => (
-                            <Table.Row key={`body-${key}-${projectName}`}>
-                                <Table.Cell>{titleCase(key)}</Table.Cell>
-                                {seqTypes.map((seq) => (
-                                    <Table.Cell key={`${key}-${seq}`}>{`${value[seq]}`}</Table.Cell>
-                                ))}
-                                <Table.Cell>
-                                    {Object.values(value).reduce((a, b) => +a + +b, 0)}
-                                </Table.Cell>
-                            </Table.Row>
-                        ))}
-                </Table.Body>
-            </Table>
-        )
-    }
-
-    const totalsStats: React.ReactElement = (
-        <>
-            <h2>Project Summary Stats</h2>
-            <b>Total Participants: </b>
-            {summary?.total_participants}
-            <br />
-            <b>Total Samples: </b>
-            {'    '}
-            {summary?.total_samples}
-            <br />
-            <b>Total Sequences: </b>
-            {'    '}
-            {summary?.total_sequences}
-            <br />
-        </>
-    )
-
-    const seqStats = () => {
-        const statsEntries = Object.entries(summary?.cram_seqr_stats ?? {})
-        if (!statsEntries.length) return <></>
-        return (
-            <Table celled>
-                <Table.Header>
-                    <Table.Row>
-                        <Table.HeaderCell>Type</Table.HeaderCell>
-                        <Table.HeaderCell>Sequences</Table.HeaderCell>
-                        <Table.HeaderCell>CRAMs</Table.HeaderCell>
-                        <Table.HeaderCell>Seqr</Table.HeaderCell>
-                    </Table.Row>
-                </Table.Header>
-                <Table.Body>
-                    {statsEntries.map(([key, value]) => (
-                        <React.Fragment key={`${key}-${projectName}`}>
-                            <Table.Row>
-                                <Table.Cell>{titleCase(key)}</Table.Cell>
-                                {Object.entries(value).map(([k1, v1]) => (
-                                    <Table.Cell key={`${key}-${k1}-${projectName}`}>
-                                        {`${v1}`}
-                                    </Table.Cell>
-                                ))}
-                            </Table.Row>
-                        </React.Fragment>
-                    ))}
-                </Table.Body>
-            </Table>
-        )
-    }
-
-    const multiQcReports = () => (
-        <>
-            <h4> MultiQC Links</h4>
-            {Object.entries(REPORT_TYPES).map(([key, value]) => (
-                <a
-                    href={`${REPORT_PREFIX}${projectName}${value}`}
-                    className="ui button"
-                    key={key}
-                    target="_blank"
-                    rel="noreferrer"
-                >
-                    {key}
-                </a>
-            ))}
-        </>
-    )
-
-    const seqrLinks = () => {
-        const seqrLinkEntries = Object.entries(summary?.seqr_links ?? {})
-        if (!seqrLinkEntries.length) return <></>
-
-        return (
-            <>
-                <h4> Seqr Links</h4>
-                {seqrLinkEntries.map(([key, value]) => (
-                    <a
-                        href={value}
-                        className="ui button"
-                        key={key}
-                        target="_blank"
-                        rel="noreferrer"
-                    >
-                        {titleCase(key)}
-                    </a>
-                ))}
-            </>
-        )
-    }
-
-    return (
-        <>
-            <br />
-            <ProjectSelector
-                setPageLimit={_setPageLimit}
-                setPageNumber={setPageNumber}
-                pageLimit={PAGE_SIZES[0]}
-            />
-            <br />
-            <hr />
-            {projectName && summary?.participants.length !== 0 && (
-                <>
-                    {totalsStats}
-                    {seqStats()}
-                    {batchTable()}
-                    <hr />
-                    {multiQcReports()}
-                    <br />
-                    <br />
-                    {seqrLinks()}
-                    <hr />
-                    <div
-                        style={{
-                            marginBottom: '10px',
-                            justifyContent: 'flex-end',
-                            display: 'flex',
-                            flexDirection: 'row',
-                        }}
-                    >
-                        <Dropdown
-                            selection
-                            onChange={setPageLimit}
-                            value={pageLimit}
-                            options={PAGE_SIZES.map((s) => ({
-                                key: s,
-                                text: `${s} samples`,
-                                value: s,
-                            }))}
-                        />
-                        {pageOptions}
-                    </div>
-                </>
-            )}
-            {renderGrid()}
-            {pageOptions}
-        </>
-    )
-}
-
-export default ProjectSummary
