+++ web/src/shared/components/links/LinkProps.ts
@@ -1,0 +1,7 @@
+import * as React from 'react'
+
+export interface LinkProps {
+    id: string
+    projectName: string
+    children: React.ReactNode
+}
