+++ web/src/shared/iconStyle.ts
@@ -1,0 +1,8 @@
+const iconStyle = {
+    fontSize: 30,
+    height: '33px',
+    verticalAlign: 'bottom',
+    marginRight: '10px',
+}
+
+export default iconStyle
